services/external/dspy_agents/requirements.txt
dspy-ai>=2.4.0
openai>=1.10.0
pydantic>=2.5.0
structlog>=24.1.0
